client: isolateserver_smoke_test

+++ client/tests/isolateserver_smoke_test.py
@@ -41,7 +41,7 @@ class IsolateServerArchiveSmokeTest(unittest.TestCase):
     # before being uploaded.
     # TODO(maruel): This should not be leaked to the client. It's a
     # transport/storage detail.
-    self.namespace = ('temporary' + str(long(time.time())).split('.', 1)[0]
+    self.namespace = ('temporary' + str(int(time.time())).split('.', 1)[0]
                       + '-gzip')
     self.tempdir = tempfile.mkdtemp(prefix=u'isolateserver')
     self.rootdir = os.path.join(self.tempdir, 'rootdir')
